corrections and rendertests
