+++ docs/scenes/disney_blend_creation/scene_light.xml
@@ -25,7 +25,7 @@
 		</film>
 	</sensor>
 
-	<bsdf type="disney" id="Material" >
+	<bsdf type="principled" id="Material" >
 		<rgb name="base_color" value="0.02, 0.02, 0.02"/>
 	</bsdf>
