+++ docs/scenes/disney_blend_creation/scene_thin_light.xml
@@ -26,9 +26,8 @@
 	</sensor>
 
 
-	<bsdf type="disney" id="Material" >
+	<bsdf type="principledthin" id="Material" >
 		<rgb name="base_color" value="1.0,1.0,1.0"/>
-        <boolean name="thin" value="true" />
 	</bsdf>
 
 	<emitter type="area" id="Background">
